+++ generatedocs/pkg/genmd/params.go
@@ -7,57 +7,16 @@ import (
 	"reflect"
 	"sort"
 	"strings"
-	"text/template"
 
 	"github.com/pkg/errors"
 	"github.com/qlik-oss/gopherciser/generatedocs/pkg/common"
 )
 
-var funcMap = template.FuncMap{
-	"join":      strings.Join,
-	"params":    handleParams,
-	"ungrouped": UngroupedActions,
-}
+const (
+	defaultIndent = "  "
+)
 
-const templateString = `{{with $data := .}}{{(index $data.Config "main").Description}}
-{{range $field, $obj := $data.ConfigFields}}{{if ne $field "scenario"}}{{with $configEntry := index $data.Config $field}}<details>
-<summary>{{$field}}</summary>
-
-{{$configEntry.Description}}
-{{params $obj}}
-{{$configEntry.Examples}}
-</details>{{end}}{{end}}{{end}}<details>
-<summary>scenario</summary>
-
-{{with $scenarioEntry := index $data.Config "scenario"}}{{$scenarioEntry.Description}}
-{{params (index $data.ConfigFields "scenario")}}{{$scenarioEntry.Examples}}
-{{range $group := $data.Groups}}<details>
-<summary>{{$group.Title}}</summary>
-
-{{$group.DocEntry.Description}}
-{{$group.DocEntry.Examples}}
-{{range $action := $group.Actions}}<details>
-<summary>{{$action}}</summary>
-{{with $actionEntry := index $data.Actions $action}}
-{{$actionEntry.Description}}
-{{with $params := params (index $data.ActionFields $action)}}### Settings
-
-{{$params}}{{end}}
-{{$actionEntry.Examples}}
-</details>{{end}}{{end}}
-</details>{{end}}{{with $groups := ungrouped}}
-<details>
-<summary>Ungrouped actions</summary>
-{{range $action := $groups}}{{with $actionEntry := index $data.Actions $action}}
-{{$actionEntry.Description}}
-{{params (index $data.ActionFields $action)}}
-{{$actionEntry.Examples}}
-</details>{{end}}{{end}}{{end}}
-{{(index $data.Extra "sessionvariables").Description}}
-</details>{{end}}
-{{(index $data.Config "main").Examples}}{{end}}`
-
-func handleParams(obj interface{}) string {
+func MarkdownParams(obj interface{}) string {
 	buf := bytes.NewBuffer(nil)
 	if err := handleValue(reflect.ValueOf(obj), buf, ""); err != nil {
 		common.Exit(err, ExitCodeFailedHandleParams)
@@ -92,7 +51,7 @@ func handleValue(value reflect.Value, buf *bytes.Buffer, indent string) error {
 			}
 
 			if value.Field(i).CanInterface() {
-				HandleFields(field, buf, indent)
+				handleFields(field, buf, indent)
 			}
 			innerIndent := indent
 			if !field.Anonymous {
@@ -140,8 +99,8 @@ func handleValue(value reflect.Value, buf *bytes.Buffer, indent string) error {
 	return nil
 }
 
-// HandleFields
-func HandleFields(field reflect.StructField, buf *bytes.Buffer, indent string) {
+// handleFields
+func handleFields(field reflect.StructField, buf *bytes.Buffer, indent string) {
 	if !field.Anonymous {
 		jsonTag := field.Tag.Get("json")
 		if jsonTag == "" {
